requirements.txt
boto3>=1.28.0
botocore>=1.31.0
pyyaml>=6.0
python-dateutil>=2.8.0
